requirements.txt
psycopg2-binary
django-redis
django
dotenv
django-environ
djangorestframework